# Conflicts:
#   src/main/java/cz/cvut/fit/corsys/pl/web/controller/LoginController.java
#   src/main/resources/application.properties

+++ src/main/java/cz/cvut/fit/corsys/dl/dao/PatientDao.java
@@ -3,6 +3,6 @@ package cz.cvut.fit.corsys.dl.dao;
 import cz.cvut.fit.corsys.dl.entity.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientDao extends JpaRepository<Patient, Long> {
+public interface PatientDao extends JpaRepository<Patient, Integer> {
 
 }
